垂直度的数据类型修改

+++ sanyou-dev-pojo/src/main/java/com/sanyou/pojo/CollectHistory.java
@@ -19,7 +19,7 @@ public class CollectHistory {
     @Column(name = "user_id")
     private String userId;
 
-    private Integer verticality;
+    private Float verticality;
 
     /**
      * @return id
@@ -80,14 +80,14 @@ public class CollectHistory {
     /**
      * @return verticality
      */
-    public Integer getVerticality() {
+    public Float getVerticality() {
         return verticality;
     }
 
     /**
      * @param verticality
      */
-    public void setVerticality(Integer verticality) {
+    public void setVerticality(Float verticality) {
         this.verticality = verticality;
     }
 }
